- Update CHANGELOG - Add tests for the etags functionality
